Moved BulletList, BulletPoint, Checkin, Collapse into new folder /dashboard
