+++ src/components/user/Dashboard.jsx
@@ -6,8 +6,8 @@ import Tile from "./Tile";
 import { QrCode, ParkingCircle } from "lucide-react";
 import Rooms from "./Rooms";
 import Packing from "./Packing";
-import BulletList from "./BulletList";
-import BulletPoints from "./BulletPoints";
+import BulletList from "./dashboard/BulletList";
+import BulletPoints from "./dashboard/BulletPoints";
 import { JUDGING } from "@/data/user/Judging";
 import { RULES } from "@/data/user/Rules";
 import Resources from "./Resources";
